+++ harness/elixir/microformats2/lib/mix/tasks/preflight-check.ex
@@ -1,0 +1,9 @@
+defmodule Mix.Tasks.PreflightCheck do
+        use Mix.Task
+
+        def run(_) do
+			Microformats2
+			Mix.Task.run("escript.build")
+        end
+end
+